Guard against NPE when using certain Contexts.
Using special Context objects, such as the BridgeContext provided by
Android Studio's previewer, would result in an NPE when calling
Mixpanel.getInstance() previously, because
ConfigurationChecker.checkBasicConfiguration() assumes that the context's
packageManager and packageName can't be null. This is now guarded
against, and Mixpanel.getInstance() will return null for contexts such as
this.