Prevent identify from overriding the ID with the same ID used in alias
Since calling identify with an aliased ID (after alias has just run)
can lead to race conditions, a check is placed so that both identify
and getPeople().identify will only change the ID if it is not the alias
which was just used. Similar logic to

https://github.com/mixpanel/mixpanel-swift/pull/49

Given that Android has 2 identify calls, to prevent identity issues,
the proposed change would make it so calling either identify would
alter the other as well.