Keep extras in the intent when tracking push open
Take out the lines that removed the extra keys (mp, mp_campaign_id and mp_message_id) from the intent when $app_open was tracked. This was originally intended to remove Mixpanel traces from the intent so it would not interfere with other libraries but this prevents multiple instances of Mixpanel from tracking the event which is the preferable outcome.